Refactor on Services Module, now Towns module
