+++ app/services/towns/cuajimalpa_de_morelos.yml
@@ -1,4 +1,4 @@
-cuajimalpa:
+cuajimalpa_de_morelos:
   1: 'col. 1ro de Mayo'
   2: 'col. Abdias García Soto</a>'
   3: 'col. Adolfo Lopez Mateos'
