+++ app/services/towns/la_magdalena_contreras.yml
@@ -1,4 +1,4 @@
-magdalena:
+la_magdalena_contreras:
   1: 'Barrio Las Calles'
   2: 'Barrio San Francisco'
   3: 'col. Acataxco'
